Refactor to introduce messageProcessor for readability.
